Improve class documentation for Applications ;
