+++ Applications/HelloRadium/minimalapp.hpp
@@ -8,41 +8,52 @@
 #include <Engine/Renderer/RenderTechnique/ShaderConfigFactory.hpp>
 
 /* This file contains a minimal radium/qt application which shows the
-classic "Spinning Cube" demo. */
+ * classic "Spinning Cube" demo.
+ */
 
-/// Our minimal application uses QTimer to be called at a regular frame rate.
+/**
+ * Our minimal application uses QTimer to be called at a regular frame rate.
+ */
 class MinimalApp : public QApplication {
     Q_OBJECT
 
   public:
-    /** IMPORTANT : the argc parameter must be a reference for calling the QApplication constructor
-        @see http://doc.qt.io/qt-5/qapplication.html#QApplication
-    */
+    /**
+     * IMPORTANT : the argc parameter must be a reference for calling the QApplication constructor
+     * @see http://doc.qt.io/qt-5/qapplication.html#QApplication
+     */
     MinimalApp( int& argc, char** argv );
 
     ~MinimalApp();
 
   public slots:
 
-    /// This function is the basic "game loop" iteration of the engine.
-    /// It starts the rendering then advance all systems by one frame.
+    /**
+     * This function is the basic "game loop" iteration of the engine.
+     * It starts the rendering then advance all systems by one frame.
+     */
     void frame();
+
+    /**
+     * Tells the application that the OpenGL context has been created.
+     * If the application needs OpenGL stuff, it has to be created here.
+     */
     void onGLInitialized();
 
   public:
-    // Our instance of the engine
+    /// Our instance of the engine.
     std::unique_ptr<Ra::Engine::RadiumEngine> m_engine;
 
-    // Task queue
+    /// Task queue.
     std::unique_ptr<Ra::Core::TaskQueue> m_task_queue;
 
-    // Pointer to Qt/OpenGL Viewer widget.
+    /// Pointer to Qt/OpenGL Viewer widget.
     std::unique_ptr<Ra::Gui::Viewer> m_viewer;
 
-    // Timer to wake us up at every frame start.
+    /// Timer to wake us up at every frame start.
     QTimer* m_frame_timer;
 
-    // Our framerate
+    /// Our framerate.
     uint m_target_fps;
 
 }; // end class
