+++ Applications/MainApplication/Gui/ColorWidget.hpp
@@ -6,6 +6,10 @@
 namespace Ra {
 namespace Gui {
 
+/**
+ * The ColorWidget class is a simple widget allowing the selection
+ * of a color when clicking it.
+ */
 class ColorWidget : public QWidget {
     Q_OBJECT
 
@@ -13,16 +17,27 @@ class ColorWidget : public QWidget {
     ColorWidget( QWidget* parent = nullptr );
 
   public slots:
+    /**
+     * Updates the widget color according to to new color.
+     */
     void colorChanged( int r, int g, int b );
 
   signals:
+    /**
+     * Emitted when a new color has been picked.
+     */
     void newColorPicked( const QColor& color );
 
   private:
+    void mousePressEvent( QMouseEvent* e ) override;
+
+    /**
+     * Updates the widget color according to to new color.
+     */
     void colorChanged();
-    virtual void mousePressEvent( QMouseEvent* e ) override;
 
   private:
+    /// The current color.
     QColor m_currentColor;
 };
