+++ Applications/MainApplication/Gui/MainWindow.hpp
@@ -36,119 +36,193 @@ class RadiumPluginInterface;
 namespace Ra {
 namespace Gui {
 
-/// This class manages most of the GUI of the application :
-/// top menu, side toolbar and side dock.
+/**
+ * This class manages most of the GUI of the application :
+ * top menu, side toolbar and side dock.
+ */
 class MainWindow : public Ra::GuiBase::MainWindowInterface, private Ui::MainWindow {
     Q_OBJECT
 
   public:
-    /// Constructor and destructor.
     explicit MainWindow( QWidget* parent = nullptr );
-    virtual ~MainWindow();
+    ~MainWindow() override;
 
-    /// Access the viewer, i.e. the rendering widget.
     Viewer* getViewer() override;
 
-    /// Access the selection manager.
     GuiBase::SelectionManager* getSelectionManager() override;
 
-    /// Update the ui from the plugins loaded.
     void updateUi( Plugins::RadiumPluginInterface* plugin ) override;
 
-    /// Update the UI ( most importantly gizmos ) to the modifications of the engine/
     void onFrameComplete() override;
 
-    // Add render in the application: UI, viewer.
     void addRenderer( std::string name, std::shared_ptr<Engine::Renderer> e ) override;
 
   public slots:
-    /// Callback to rebuild the item model when the engine objects change.
-    void onItemAdded( const Engine::ItemEntry& ent );
-
-    void onItemRemoved( const Engine::ItemEntry& ent );
-
-    // Frame timers ui slots
-    void onUpdateFramestats( const std::vector<FrameTimerData>& stats );
-
-    // Selection tools
-    void onSelectionChanged( const QItemSelection& selected, const QItemSelection& deselected );
-
-    // Gizmo buttons slots
-    void gizmoShowNone();
-    void gizmoShowTranslate();
-    void gizmoShowRotate();
-    void gizmoShowScale();
-
-    // Keymapping configuration tool
-    void reloadConfiguration();
-    void loadConfiguration();
-
-    /// Slot for the "visible" button
-    void toggleVisisbleRO();
-
-    /// Reset the camera to see all visible objects
-    void fitCamera();
     void postLoadFile() override { fitCamera(); }
 
-    /// Slot for the "edit" button.
-    void editRO();
-
-    /// Cleanup resources.
     void cleanup() override;
 
+    /**
+     * Callback to rebuild the item model when adding an object to the engine.
+     */
+    void onItemAdded( const Engine::ItemEntry& ent );
+
+    /**
+     * Callback to rebuild the item model when removing an object from the engine.
+     */
+    void onItemRemoved( const Engine::ItemEntry& ent );
+
+    /**
+     * Updates the frame timers ui.
+     */
+    void onUpdateFramestats( const std::vector<FrameTimerData>& stats );
+
+    /**
+     * Manages the selection tools.
+     */
+    void onSelectionChanged( const QItemSelection& selected, const QItemSelection& deselected );
+
+    /// \name Gizmo tools
+    /// \{
+
+    /**
+     * Toggle on/off displaying gizmos.
+     */
+    void gizmoShowNone();
+
+    /**
+     * Use the translation gizmo.
+     */
+    void gizmoShowTranslate();
+
+    /**
+     * Use the ratotion gizmo.
+     */
+    void gizmoShowRotate();
+
+    /**
+     * Use the scaling gizmo.
+     */
+    void gizmoShowScale();
+    /// \}
+
+    /// \name Keymapping configuration tool
+    /// \{
+
+    /**
+     * Reloads the keymapping configuration from file, if it exists.
+     */
+    void reloadConfiguration();
+
+    /**
+     * Loads the keymapping configuration from a file.
+     */
+    void loadConfiguration();
+    /// \}
+
+    /**
+     * Toggle on/off displaying the selected object.
+     */
+    void toggleVisisbleRO();
+
+    /**
+     * Reset the camera to see all visible objects.
+     */
+    void fitCamera();
+
+    /**
+     * Use the current selected RenderObject, if any, for material edition.
+     */
+    void editRO();
+
   signals:
-    /// Emitted when the frame loads
+    /**
+     * Emitted when the loading a file.
+     */
     void fileLoading( const QString path );
 
-    /// Emitted when the user changes the timer box ("Frame average over count")
+    /**
+     * Emitted when the user changes the timer box ("Frame average over count").
+     */
     void framescountForStatsChanged( int count );
 
-    /// Emitted when a new item is selected. An invalid entry is sent when no item is selected.
+    /**
+     * Emitted when a new item is selected. An invalid entry is sent when no item is selected.
+     */
     void selectedItem( const Engine::ItemEntry& entry );
 
   private:
-    /// Connect qt signals and slots. Called once by the constructor.
+    void closeEvent( QCloseEvent* event ) override;
+
+    /**
+     * Connect qt signals and slots. Called once by the constructor.
+     */
     void createConnections();
 
-    // Updates the Tracking info according to the selected feature, if any.
+    /**
+     * Updates the Tracking info according to the selected feature, if any.
+     */
     void updateTrackedFeatureInfo();
 
-    virtual void closeEvent( QCloseEvent* event ) override;
-
-    /// Update displayed texture according to the current renderer
+    /**
+     * Update displayed texture according to the current renderer.
+     */
     void updateDisplayedTexture();
 
   private slots:
-    /// Slot for the "load file" menu.
+    /**
+     * Slot for the "load file" menu.
+     */
     void loadFile();
 
-    /// Slot for the "material editor"
+    /**
+     * Slot for the "material editor".
+     */
     void openMaterialEditor();
 
-    /// Slot for the user changing the current shader
+    /**
+     * Slot for the user changing the current shader.
+     */
     void changeRenderObjectShader( const QString& shaderName );
 
-    /// Slot for the user changing the current renderer
+    /**
+     * Slot for the user changing the current renderer.
+     */
     void onCurrentRenderChangedInUI();
 
-    /// Slot for the picking results from the viewer.
+    /**
+     * Slot for the picking results from the viewer.
+     */
     void handlePicking( const Ra::Engine::Renderer::PickingResult& pickingResult );
 
-    // Slot to init renderers once gl is ready
+    /**
+     * Slot to init renderers once gl is ready.
+     */
     void onGLInitialized();
 
-    /// Slot to accept a new renderer
+    /**
+     * Slot to accept a new renderer.
+     */
     void onRendererReady();
 
-    /// Exports the mesh of the currently selected object to a file.
+    /**
+     * Exports the mesh of the currently selected object to a file.
+     */
     void exportCurrentMesh();
 
-    /// Remove the currently selected item (entity, component or ro)
+    /**
+     * Remove the currently selected item (entity, component or ro).
+     */
     void deleteCurrentItem();
 
-    /// Clears all entities and resets the camera.
+    /**
+     * Clears all entities and resets the camera.
+     */
     void resetScene();
 
+    /**
+     * Toggle on/off displaying the picking circle.
+     */
     void toggleCirclePicking( bool on );
 
   private:
@@ -161,7 +235,7 @@ class MainWindow : public Ra::GuiBase::MainWindowInterface, private Ui::MainWind
     /// Widget to allow material edition.
     MaterialEditor* m_materialEditor;
 
-    /// viewer widget
+    /// Viewer widget.
     Ra::Gui::Viewer* m_viewer;
 };
